- удалил дублирующие классы для листа вакансий и деталей (наприме Salary) - рефакторинг кода для ci - где необходимо добавил аннотацию @SerializedName - удалил ненужные поля в класса - перераспределил классы по пакетам
