--- app/src/main/java/ru/practicum/android/diploma/data/dto/responses/deteils_vacancy/Area.kt
@@ -1,6 +1,0 @@
-package com.example.myapplication.models.current_vacancy
-
-data class Area(
-    val id: String,
-    val name: String, // название региона
-)
