--- app/src/main/java/ru/practicum/android/diploma/data/dto/responses/deteils_vacancy/Employer.kt
@@ -1,8 +1,0 @@
-package com.example.myapplication.models.current_vacancy
-
-data class Employer(
-    val id: String,
-    val logo_urls: LogoUrls, // логотипы разных размеров компании
-    val name: String, // название компании
-    val trusted: Boolean, // прошлали компания проверку на сайте??
-)
