--- app/src/main/java/ru/practicum/android/diploma/data/dto/responses/deteils_vacancy/KeySkill.kt
@@ -1,5 +1,0 @@
-package com.example.myapplication.models.current_vacancy
-
-data class KeySkill(
-    val name: String
-)
