+++ app/src/main/java/ru/practicum/android/diploma/data/dto/responses/guides/areas/AreaGuide.kt
@@ -1,0 +1,8 @@
+package ru.practicum.android.diploma.data.dto.responses.guides.areas
+
+data class AreaGuide(
+    val areas: List<AreaGuide>?,
+    val id: String,
+    val name: String,
+    val parent_id: String
+)
