+++ app/src/main/java/ru/practicum/android/diploma/data/dto/responses/guides/areas/ResponseAreasGuide.kt
@@ -1,0 +1,5 @@
+package ru.practicum.android.diploma.data.dto.responses.guides.areas
+
+import ru.practicum.android.diploma.data.dto.responses.Response
+
+data class ResponseAreasGuide(val listItem: ArrayList<ResponseAreasGuideItem>) : Response()
