+++ app/src/main/java/ru/practicum/android/diploma/data/dto/responses/guides/areas/ResponseAreasGuideItem.kt
@@ -1,7 +1,7 @@
-package com.example.myapplication.models.model_guide_area
+package ru.practicum.android.diploma.data.dto.responses.guides.areas
 
 data class ResponseAreasGuideItem(
-    val areas: List<Area>, // список локаций
+    val areas: List<AreaGuide>, // список локаций
     val id: String,
     val name: String, // название локации
     val parent_id: Any // id родительской локации
