+++ app/src/main/java/ru/practicum/android/diploma/data/dto/responses/guides/inustries/Industry.kt
@@ -1,4 +1,4 @@
-package com.example.myapplication.models.model_guide_inustries
+package ru.practicum.android.diploma.data.dto.responses.guides.inustries
 
 data class Industry(
     val id: String,
