+++ app/src/main/java/ru/practicum/android/diploma/data/dto/responses/guides/inustries/ResponseGuideIndustriesDto.kt
@@ -1,4 +1,4 @@
-package com.example.myapplication.models.model_guide_inustries
+package ru.practicum.android.diploma.data.dto.responses.guides.inustries
 
 import ru.practicum.android.diploma.data.dto.responses.Response
