--- app/src/main/java/ru/practicum/android/diploma/data/dto/responses/list_vacancy/LogoUrls.kt
@@ -1,7 +1,0 @@
-package com.example.myapplication.models.list_vacancy
-
-data class LogoUrls(
-    val `240`: String,
-    val `90`: String,
-    val original: String
-)
