--- app/src/main/java/ru/practicum/android/diploma/data/dto/responses/model_guide_area/Area.kt
@@ -1,8 +1,0 @@
-package com.example.myapplication.models.model_guide_area
-
-data class Area(
-    val areas: List<Area>?,
-    val id: String,
-    val name: String,
-    val parent_id: String
-)
