--- app/src/main/java/ru/practicum/android/diploma/data/dto/responses/model_guide_area/ResponseAreasGuide.kt
@@ -1,5 +1,0 @@
-package com.example.myapplication.models.model_guide_area
-
-import ru.practicum.android.diploma.data.dto.responses.Response
-
-data class ResponseAreasGuide(val listItem: ArrayList<ResponseAreasGuideItem>): Response()
