+++ app/src/main/java/ru/practicum/android/diploma/data/dto/responses/vacancy/Area.kt
@@ -1,4 +1,4 @@
-package com.example.myapplication.models.list_vacancy
+package ru.practicum.android.diploma.data.dto.responses.vacancy
 
 data class Area(
     val name: String, // название города
