+++ app/src/main/java/ru/practicum/android/diploma/data/dto/responses/vacancy/deteils/Experience.kt
@@ -1,4 +1,4 @@
-package com.example.myapplication.models.current_vacancy
+package ru.practicum.android.diploma.data.dto.responses.vacancy.deteils
 
 data class Experience(
     val id: String,
