+++ app/src/main/java/ru/practicum/android/diploma/data/dto/responses/vacancy/deteils/KeySkill.kt
@@ -1,0 +1,5 @@
+package ru.practicum.android.diploma.data.dto.responses.vacancy.deteils
+
+data class KeySkill(
+    val name: String
+)
