+++ app/src/main/java/ru/practicum/android/diploma/data/dto/responses/vacancy/list/ResponseListDto.kt
@@ -1,11 +1,12 @@
-package com.example.myapplication.models.list_vacancy
+package ru.practicum.android.diploma.data.dto.responses.vacancy.list
 
+import com.google.gson.annotations.SerializedName
 import ru.practicum.android.diploma.data.dto.responses.Response
 
-data class ResponseListVacancyDto(
+data class ResponseListDto(
     val found: Int,  // всего найдено по запросу
     val items: List<Item>, // список вакансий
     val page: Int, // номер страницы
     val pages: Int, // всего страниц
-    val per_page: Int, // кол-во итемов на странице
+    @SerializedName("per_page") val perPage: Int, // кол-во итемов на странице
 ) : Response()
